Apply Clippy Lints for MerkleProof
